+++ changelog.md
@@ -1,9 +1,23 @@
 ﻿# Changelog
 
+v0.0.2.4
+ - added class RecordingChannel, will present a channel in the RecordingItem class
+   - added methode to give back a List of RecordingChannel to get all channels in the recordings
+ - RecordingItem
+   - renamed field SDuration to Duration2 and changed type from string to TimeSpan
+   - note: Duration gives the unformated string back, thats the media server sends
+   - the field RecDate gives back a DateTime from the unformated field StartDate
+   - the field Channel is now a object of type RecordingChannel
+ - added class VideoFilePath
+ - added class VideoFilePathItem
+   - every PathItem presents a Path from the videofolder
+   - every PathItem contains his own Parent and his Childfolders
+   - every PathItem gives back his owned videofiles or all videos in all subdirectories
+
 v0.0.2.3
 - changed access modifiers of getter methods in some classes
 - renamed Recording to RecordingList
-- add RecordingSeries as seperate object to RecordingItem
+- added field RecordingSeries as seperate object to RecordingItem
 - Recordings can now be filtered by Series
 - added functions to generate M3U Files for video and recordings
 - some improvements and bugfixes